# src/main/java/org/littlered/dataservices/rest/params/eventManager/AddUserRole.java
package org.littlered.dataservices.rest.params.eventManager;

public class AddUserRole {

	private Long userId;
	private String role;

	public Long getUserId() {
		return userId;
	}

	public void setUserId(Long userId) {
		this.userId = userId;
	}

	public String getRole() {
		return role;
	}

	public void setRole(String password) {
		this.role = password;
	}
}