# src/main/java/org/littlered/dataservices/rest/params/eventManager/ConfirmPasswordRequest.java
package org.littlered.dataservices.rest.params.eventManager;

public class ConfirmPasswordRequest {

	private String token;

	private String password;

	public String getToken() {
		return token;
	}

	public void setToken(String token) {
		this.token = token;
	}

	public String getPassword() {
		return password;
	}

	public void setPassword(String password) {
		this.password = password;
	}
}