# src/main/java/org/littlered/dataservices/rest/params/eventManager/PasswordResetRequest.java
package org.littlered.dataservices.rest.params.eventManager;

public class PasswordResetRequest {

	private String email;

	public String getEmail() {
		return email;
	}

	public void setEmail(String email) {
		this.email = email;
	}
}