# src/main/java/org/littlered/dataservices/rest/params/eventManager/UserPassword.java
package org.littlered.dataservices.rest.params.eventManager;

public class UserPassword {

	private Long userId;
	private String password;

	public Long getUserId() {
		return userId;
	}

	public void setUserId(Long userId) {
		this.userId = userId;
	}

	public String getPassword() {
		return password;
	}

	public void setPassword(String password) {
		this.password = password;
	}
}