# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API for Big Bad Con",
    "version": "1.0",
    "title": "Big Bad Con data service REST API",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Jeremy Tidwell",
      "url": "https://www.bigbadcon.com/",
      "email": "necanthrope@gmail.com"
    },
    "license": {
      "name": "Apache License Version 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8091",
  "basePath": "/",
...
  "paths": {
...
    "/users/addRoleToUser": {
      "post": {
        "tags": [
          "users-controller"
        ],
        "summary": "Add a role to a user. Admin only.",
        "operationId": "addRoleToUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userRole",
            "description": "userRole",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUserRole"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/all": {
      "get": {
        "tags": [
          "users-controller"
        ],
        "summary": "Return all registered users. Admin only.",
        "operationId": "getAllUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/confirmPasswordRequest": {
      "post": {
        "tags": [
          "users-controller"
        ],
        "summary": "confirmPasswordRequest",
        "operationId": "confirmPasswordRequestUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "request",
            "description": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ConfirmPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/create": {
      "put": {
        "tags": [
          "users-controller"
        ],
        "summary": "Create a new user.",
        "operationId": "createUserUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "createUser",
            "description": "createUser",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUsersDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/id/{id}": {
      "get": {
        "tags": [
          "users-controller"
        ],
        "summary": "Display information on a user, searching by ID. Admin only.",
        "operationId": "findUserByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UsersDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "users-controller"
        ],
        "summary": "Display information on the current user.",
        "operationId": "findMeUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UsersDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/me/isadmin": {
      "get": {
        "tags": [
          "users-controller"
        ],
        "summary": "Indicate if the current user has admin rights.",
        "operationId": "isAdminUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/resetPasswordRequest": {
      "post": {
        "tags": [
          "users-controller"
        ],
        "summary": "resetPasswordRequest",
        "operationId": "resetPasswordRequestUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "request",
            "description": "request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PasswordResetRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/setMyPassword": {
      "post": {
        "tags": [
          "users-controller"
        ],
        "summary": "Set password for currently logged in user. Ignores userId field in input.",
        "operationId": "setPasswordForMeUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userPassword",
            "description": "userPassword",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserPassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/setPassword": {
      "post": {
        "tags": [
          "users-controller"
        ],
        "summary": "Set password for a user. Admin only.",
        "operationId": "setPasswordForUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userPassword",
            "description": "userPassword",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserPassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/username/{username}": {
      "get": {
        "tags": [
          "users-controller"
        ],
        "summary": "Display information on a user, searching by username. Admin only.",
        "operationId": "findUserByUsernameUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "username",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UsersDTO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
...
    "ConfirmPasswordRequest": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    },
...
    "Timestamp": {
      "type": "object",
      "properties": {
        "date": {
          "type": "integer",
          "format": "int32"
        },
        "day": {
          "type": "integer",
          "format": "int32"
        },
        "hours": {
          "type": "integer",
          "format": "int32"
        },
        "minutes": {
          "type": "integer",
          "format": "int32"
        },
        "month": {
          "type": "integer",
          "format": "int32"
        },
        "nanos": {
          "type": "integer",
          "format": "int32"
        },
        "seconds": {
          "type": "integer",
          "format": "int32"
        },
        "time": {
          "type": "integer",
          "format": "int64"
        },
        "timezoneOffset": {
          "type": "integer",
          "format": "int32"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "PasswordResetRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        }
      }
    },
    "UsersDTO": {
      "type": "object",
      "properties": {
        "bbcUserFavorites": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BbcUserFavorites"
          }
        },
        "bookingComment": {
          "type": "string"
        },
        "displayName": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "lastUpdated": {
          "type": "integer",
          "format": "int64"
        },
        "metadata": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Usermeta"
          }
        },
        "userEmail": {
          "type": "string"
        },
        "userNicename": {
          "type": "string"
        },
        "userRegistered": {
          "$ref": "#/definitions/Timestamp"
        },
        "userUrl": {
          "type": "string"
        }
      }
    },
...
    "CreateUsersDTO": {
      "type": "object",
      "properties": {
        "displayName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "userEmail": {
          "type": "string"
        },
        "userLogin": {
          "type": "string"
        },
        "userNicename": {
          "type": "string"
        },
        "userPass": {
          "type": "string"
        },
        "userUrl": {
          "type": "string"
        }
      }
    },
...
    "UserPassword": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "BbcUserFavorites": {
      "type": "object",
      "properties": {
        "bbcUserFavoritesId": {
          "type": "integer",
          "format": "int64"
        },
        "createDate": {
          "$ref": "#/definitions/Timestamp"
        },
        "eventId": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "updateDate": {
          "$ref": "#/definitions/Timestamp"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "Usermeta": {
      "type": "object",
      "properties": {
        "metaKey": {
          "type": "string"
        },
        "metaValue": {
          "type": "string"
        },
        "umetaId": {
          "type": "integer",
          "format": "int64"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "AddUserRole": {
      "type": "object",
      "properties": {
        "role": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
  }
}